+++ tests/test_matrices.py
@@ -617,54 +617,6 @@ def test_indexing_ix_both(mat):
     np.testing.assert_array_equal(res, expected)
 
 
-def test_pandas_to_matrix():
-    n_rows = 50
-    dense_column = np.linspace(-10, 10, num=n_rows, dtype=np.float64)
-    dense_column_with_lots_of_zeros = dense_column.copy()
-    dense_column_with_lots_of_zeros[:44] = 0.0
-    sparse_column = np.zeros(n_rows, dtype=np.float64)
-    sparse_column[0] = 1.0
-    cat_column_lowdim = np.tile(["a", "b"], n_rows // 2)
-    cat_column_highdim = np.arange(n_rows)
-
-    dense_ser = pd.Series(dense_column)
-    lowdense_ser = pd.Series(dense_column_with_lots_of_zeros)
-    sparse_ser = pd.Series(sparse_column, dtype=pd.SparseDtype("float", 0.0))
-    cat_ser_lowdim = pd.Categorical(cat_column_lowdim)
-    cat_ser_highdim = pd.Categorical(cat_column_highdim)
-
-    df = pd.DataFrame(
-        data={
-            "d": dense_ser,
-            "ds": lowdense_ser,
-            "s": sparse_ser,
-            "cl_obj": cat_ser_lowdim.astype(object),
-            "ch": cat_ser_highdim,
-        }
-    )
-
-    mat = tm.from_pandas(
-        df, dtype=np.float64, sparse_threshold=0.3, cat_threshold=4, object_as_cat=True
-    )
-
-    assert mat.shape == (n_rows, n_rows + 5)
-    assert len(mat.matrices) == 3
-    assert isinstance(mat, tm.SplitMatrix)
-
-    nb_col_by_type = {
-        tm.DenseMatrix: 3,  # includes low-dimension categorical
-        tm.SparseMatrix: 2,  # sparse column
-        tm.CategoricalMatrix: n_rows,
-    }
-    for submat in mat.matrices:
-        assert submat.shape[1] == nb_col_by_type[type(submat)]
-
-    # Prevent a regression where the column type of sparsified dense columns
-    # was being changed in place.
-    assert df["cl_obj"].dtype == object
-    assert df["ds"].dtype == np.float64
-
-
 @pytest.mark.parametrize("mat", get_all_matrix_base_subclass_mats())
 def test_split_matrix_creation(mat):
     sm = tm.SplitMatrix(matrices=[mat, mat])
@@ -839,61 +791,3 @@ def test_combine_names(mat_1, mat_2):
 
     assert combined.column_names == mat_1.column_names + mat_2.column_names
     assert combined.term_names == mat_1.term_names + mat_2.term_names
-
-
-@pytest.mark.parametrize("prefix_sep", ["_", ": "])
-@pytest.mark.parametrize("drop_first", [True, False])
-def test_names_pandas(prefix_sep, drop_first):
-    n_rows = 50
-    dense_column = np.linspace(-10, 10, num=n_rows, dtype=np.float64)
-    dense_column_with_lots_of_zeros = dense_column.copy()
-    dense_column_with_lots_of_zeros[:44] = 0.0
-    sparse_column = np.zeros(n_rows, dtype=np.float64)
-    sparse_column[0] = 1.0
-    cat_column_lowdim = np.tile(["a", "b"], n_rows // 2)
-    cat_column_highdim = np.arange(n_rows)
-
-    dense_ser = pd.Series(dense_column)
-    lowdense_ser = pd.Series(dense_column_with_lots_of_zeros)
-    sparse_ser = pd.Series(sparse_column, dtype=pd.SparseDtype("float", 0.0))
-    cat_ser_lowdim = pd.Categorical(cat_column_lowdim)
-    cat_ser_highdim = pd.Categorical(cat_column_highdim)
-
-    df = pd.DataFrame(
-        data={
-            "d": dense_ser,
-            "cl_obj": cat_ser_lowdim.astype(object),
-            "ch": cat_ser_highdim,
-            "ds": lowdense_ser,
-            "s": sparse_ser,
-        }
-    )
-
-    categorical_format = "{name}" + prefix_sep + "{category}"
-    mat_end = tm.from_pandas(
-        df,
-        dtype=np.float64,
-        sparse_threshold=0.3,
-        cat_threshold=4,
-        object_as_cat=True,
-        cat_position="end",
-        categorical_format=categorical_format,
-        drop_first=drop_first,
-    )
-
-    expanded_df = pd.get_dummies(df, prefix_sep=prefix_sep, drop_first=drop_first)
-    assert mat_end.column_names == expanded_df.columns.tolist()
-
-    mat_expand = tm.from_pandas(
-        df,
-        dtype=np.float64,
-        sparse_threshold=0.3,
-        cat_threshold=4,
-        object_as_cat=True,
-        cat_position="expand",
-        categorical_format=categorical_format,
-        drop_first=drop_first,
-    )
-
-    unique_terms = list(dict.fromkeys(mat_expand.term_names))
-    assert unique_terms == df.columns.tolist()
