+++ tests/test_split_matrix.py
@@ -306,17 +306,3 @@ def test_matvec(n_rows):
     )
     mat = from_pandas(X, cat_threshold=0)
     np.testing.assert_allclose(mat.matvec(np.array(mat.shape[1] * [1])), n_cols)
-
-
-@pytest.mark.parametrize("cat_missing_method", ["fail", "zero", "convert"])
-def test_from_pandas_missing(cat_missing_method):
-    df = pd.DataFrame({"cat": pd.Categorical([1, 2, pd.NA, 1, 2, pd.NA])})
-    if cat_missing_method == "fail":
-        with pytest.raises(
-            ValueError, match="Categorical data can't have missing values"
-        ):
-            from_pandas(df, cat_missing_method=cat_missing_method)
-    elif cat_missing_method == "zero":
-        assert from_pandas(df, cat_missing_method=cat_missing_method).shape == (6, 2)
-    elif cat_missing_method == "convert":
-        assert from_pandas(df, cat_missing_method=cat_missing_method).shape == (6, 3)
